% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{predictor}
\alias{predictor}
\title{A C50 Predictor Function}
\usage{
predictor(file = "data.csv", cnt = 4)
}
\arguments{
\item{data.csv}{and counter = 4}
}
\description{
Suspect(homicide) probaility predictor using C50 (decision tree variant) algorithm.
}
\keyword{C50}
\keyword{Predictor}
